Refactor hedge calculator into modular sections
